apps/api/pyproject.toml
[project]
name = "nexus-api"
version = "0.1.0"
description = "NEXUS AI — multi-tenant SaaS backend (FastAPI)"
authors = [{ name = "ShonazarXudayberganov", email = "shonazarx4@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.0,<0.116.0)",
    "uvicorn[standard] (>=0.32.0,<0.33.0)",
    "pydantic (>=2.9.0,<3.0.0)",
    "pydantic-settings (>=2.6.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.36,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "redis (>=5.2.0,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "bcrypt (>=4.2.0,<5.0.0)",
    "python-multipart (>=0.0.17,<0.1.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "phonenumbers (>=8.13.0,<9.0.0)",
    "structlog (>=24.4.0,<25.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
types-python-jose = "^3.3.4"
pre-commit = "^4.0.0"
faker = "^33.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

# ─────────── Ruff ───────────
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "ASYNC",# flake8-async
    "S",    # flake8-bandit (security)
    "RUF",  # ruff-specific
]
ignore = [
    "S101",  # use of assert (ok in tests)
    "B008",  # do not perform function calls in argument defaults (FastAPI Depends)
    "S105",  # hardcoded "Bearer" / "access" / "refresh" string false positives
    "RUF012",  # SQLAlchemy __table_args__ pattern is conventional, not mutable state
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S107"]
"alembic/env.py" = ["E402"]
"app/core/db.py" = ["BLE001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─────────── Mypy ───────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# ─────────── Pytest ───────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (DB, Redis)",
    "e2e: End-to-end tests (full stack)",
]

[tool.coverage.run]
source = ["app"]
omit = ["app/main.py", "alembic/*", "tests/*"]
concurrency = ["greenlet", "thread"]
